♻️ refactor(composio): improve module separation and type organization
- Move ConnectorType to shared types module for better reusability
- Relocate validateEnvironment function to server-side only module
- Update import paths to use centralized type definitions
- Remove client-side validateEnvironment from utils (server-side only)
- Improve server/client boundary separation for better security
