+++ lib/composio-utils.ts
@@ -1,11 +1,7 @@
 // Client-safe utility functions for Composio connectors
 // This file contains only utility functions that can be safely used on the client-side
 
-export type ConnectorType =
-  | 'gmail'
-  | 'googlecalendar'
-  | 'notion'
-  | 'googledrive';
+import type { ConnectorType } from './types';
 
 // Map connector types to auth config IDs using public environment variables
 export const getAuthConfigId = (connectorType: ConnectorType): string => {
@@ -29,19 +25,4 @@ export const getAuthConfigId = (connectorType: ConnectorType): string => {
   }
 };
 
-/**
- * Validate environment setup (can be used server-side only due to non-public env var)
- */
-export const validateEnvironment = (): {
-  isValid: boolean;
-  message?: string;
-} => {
-  if (!process.env.COMPOSIO_API_KEY) {
-    return {
-      isValid: false,
-      message: 'COMPOSIO_API_KEY environment variable is not set',
-    };
-  }
-
-  return { isValid: true };
-};
+// validateEnvironment moved to composio-server.ts (server-side only)
